Zero-extend the per-shape term in file_load_shape2d_expandedsize
file_load_shape2d_expandedsize (asmorig/seg037.asm:50) accumulates into a
32-bit total and is explicit about how each per-shape term is widened:

        mov     ax, es:[bx+SHAPE2D.s2d_height]
        imul    es:[bx+SHAPE2D.s2d_width]
        mov     cl, 3
        shl     ax, cl              ; 16-bit, carries dropped
        mov     [bp+var_6], ax
        sub     dx, dx              ; high half is ZERO, not a sign extension
        add     [bp+var_sizelo], ax
        adc     [bp+var_sizehi], dx
        add     [bp+var_sizelo], size SHAPE2D
        adc     [bp+var_sizehi], 0

The C port wrote the whole term as one 16-bit int expression and let the
conversion to long sign-extend it, so any shape whose w*h*8 came out with bit
15 set subtracted from the total instead of adding to it. This has to agree
with the offsets file_load_shape2d_expand lays down, since this function
sizes the chunk that function then fills.

Cast through unsigned short and add the header size separately, matching the
two adds. The seed value stays as it is: the original really does sign-extend
that one (`shl ax, cl / add ax, size SHAPE2D / cwd`).

+++ src/restunts/c/shape2d.c
@@ -517,7 +517,11 @@ unsigned short file_load_shape2d_expandedsize(void far* memchunk) {
 
 	for (i = 0; i < shapecount; ++i) {
 		memshape = file_get_shape2d(memchunk, i);
-		size += (memshape->s2d_width * memshape->s2d_height * 8) + sizeof(struct SHAPE2D);
+		// `shl ax, 3` then `sub dx, dx / adc`: the per-shape term is a
+		// 16-bit value ZERO-extended into the accumulator, and the header
+		// size is folded in afterwards with its own carry.
+		size += (unsigned long)(unsigned short)(memshape->s2d_width * memshape->s2d_height * 8)
+		      + sizeof(struct SHAPE2D);
 	}
 
 	return (size + sizeof(struct SHAPE2D)) >> 4;
